[#2114] Move route tests to mirror app structure
Move test files to src/tests/apps/ to match route file locations:
- src/tests/apps/secret/routes/ - metadata.spec.ts, secret.spec.ts
- src/tests/apps/session/routes.spec.ts
- src/tests/apps/workspace/routes/ - account.spec.ts, dashboard.spec.ts

Update imports to use new route paths.
